Edit user implementation_client

+++ client/src/api/index.js
@@ -3,12 +3,23 @@ import axios from 'axios'
 
 const API = axios.create({ baseURL: 'http://localhost:3001'})
 
+const profile = JSON.parse(localStorage.getItem('profile'))
+
+const accessHeader = {
+    headers: {
+      'Authorization': `token ${profile ? profile.accessToken : null}`
+    }
+  }
+
 export const loginIn = (formData) => API.post('/users/v1/login', formData)
 
 export const register = (formData) => API.post('/users/v1/register', formData)
 
-export const deleteUser = (formData) => API.delete('/users/v1/delete', {data:formData})
+export const deleteUser = (formData) => API.delete('/users/v1/delete', {data:formData}, accessHeader)
 
-export const updatePassword = (formData) =>API.post('/users/v1/updatePassword', formData)
+export const updatePassword = (formData) =>API.post('/users/v1/updatePassword', formData, accessHeader)
 
-export const getUser = (formData) => API.post('/users/v1/view', formData)
+export const getUser = (formData) => API.post('/users/v1/view', formData, accessHeader)
+
+export const editUser = (formData) => API.post('/users/v1/edit', formData, accessHeader)
+
